requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
mlxtend
sodapy
streamlit
plotly.express
folium
streamlit_folium